+++ core/vgui/internals/vgui_rasterpos.cxx
@@ -43,7 +43,7 @@ void vgui_rasterpos4dv(double const X[4])
   glRasterPos2f(0,0);
   glBitmap(0,0,
            0,0,
-           rx*vp[2]/2, ry*vp[3]/2,
+           float(rx*vp[2]/2), float(ry*vp[3]/2),
            0);
 }
